layout/forms/test/test_bug869314.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=869314
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 869314</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <style type="text/css">
    .selectbox {
      background-color: #00FF00;
    }
  </style>

</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=869314">Mozilla Bug 869314</a>
<p id="display"></p>
<div id="content">

  <select id="selectbox1" name="non-native selectbox" class="selectbox">
    <option value="item">test item</option>
  </select>

  <select id="selectbox2" name="native selectbox">
    <option value="item">test item</option>
  </select>

  <script type="application/javascript">
  ok(document.getElementById("selectbox1").clientWidth >
     document.getElementById("selectbox2").clientWidth,
     "Non-native styled combobox does not have enough space for a dropmarker!");
  </script>

</div>
<pre id="test">
</pre>
</body>
</html>
